Rewrites how allX and unlockedX arrays are created
There was originally very little connection between the names of the 'allX' arrays and 'unlockedX' arrays.
Similarly, 'unlockedWeapons' and other arrays like 'unlockedX' risked falling out of sync, given everything in 'unlockedWeapons' should also be in another array.

To begin to mitigate this, now all of the different 'categories' of variables are written out, and the 'allX' and 'unlockedX' variables are created to match those categories.

Right now, 'special' categories aren't included, as they require a lot of additional work to identify, except for 'AA', 'AT', and 'GrenadeLaunchers' - because the existing code uses these for 'unlockedX' arrays.

This lays the groundwork for unifying the unlocking under one system (to avoid arrays falling out of sync).