Prevent hardcoded city pop data from overwriting processed data (#713)
Moved the hardcoded city population data to initZones, which is the only place that it's used. Previously it was being initialized on clients & HCs. Since recent init order changes, it's been overwriting the city data from the server that's stored in the same logic object.

initZones is a garbage fire, and reusing the same keys for completely different data is nasty, but this will fix problems for the moment.

Also adjusted initZones to use the hardcoded pop data for Chernarus Winter, matching the behaviour for Chernarus Summer.