Speculative workaround for remaining schrodinger's buildings (#1010)
A speculative workaround for a potential Arma bug with the synchronization of destroyed buildings. Currently we JIP-remoteExec hideObject for each building that was destroyed at server start, and we have an additional fix (fn_installSchrodingersBuildingFix) for the case where a building is destroyed while a client is connected but not correctly updated on all clients.

This PR covers a third case where a building is destroyed after server load but before a specific client joins. The current destroyedBuildings list is sent to each client on connection, and the client then hides those buildings. The list is sent using a request/send system to avoid unnecessary publicVariable spam on building destruction. As the remoteExec count is much lower, this may also dodge a potential issue where the hideObject remoteExecs were lost due to network spam