Save overhaul, mostly for player saves:
- Fixes bug where player saves wouldn't load correctly due to multiple global setVariables.
- Fixes bug where player saving could be enabled slightly before their stats were loaded.
- Fixes bug where autosave would spam additional player saves.
- Fixes minor exploit where players could gain money by not loading their previous game.
- Fixes issue where autosave wouldn't trigger unless there was a commander on the server.
- Fixes bug where player loadout wasn't loaded in SP.
- Works around unknown-cause issue where player unit lacked variables on disconnection.
- Adds some save/load logging to diagnose future broken player saves.

The first bug needed a substantial reorganisation so that the global variables related to the
player save are always set globally from the server on startup. There's now a resetPlayer function
which handles the case of selecting "no" to load previous session. This function sets the player
money to the minimum of 100 and the player's previous money, which removes the need for setting
the startup money to 95.