Tweak QRF balance (#1311)
### What have you changed and why?
Various issues related to QRF balance:

#### Limited QRF spawning close to rebels
Prevented QRF spawning closer than half spawn distance from rebels. Outposts and airports beyond spawn distance are preferred. Note that the check applies to the the marker, not the spawn point, so airports may be a bit weird.

#### Increased QRF delay
Increased the delay between QRFs to/near the same marker to 20-30 mins from 10-20 as requested.

#### Softened the aggro -> vehicle count relationship
I thought this was over-adjusted given that QRFs are also being multiplied by AOs, and the constant one-vehicle QRFs in SP are pitiful. Replaced some aggro adjustment with a bit of randomness.  The range is now more like 1.25 to 5 vehicles rather than 0-6 min 1.

Should be re-examined if we stop multiplying QRFs by AOs, but it's not a strong change anyway.

#### Rebalanced QRF vehicle selection
- Advanced vehicles are available earlier, probabilities increased more gradually.
- Basic vehicles retained at higher war levels but at low probabilities.
- Proportion of squad-carrying vehicles increased, especially at higher war levels.
- Attempt to balance air vs land removed, as this is not how the function is used.

The method still has the old flaws, although these should be mitigated somewhat by using a wider vehicle selection:
- There's no reasonable way to adjust for missing vehicle categories, either due to the modset (AFRF lacking air transports) or due to economicsAI (eg. government running out of APCs).
- Single available vehicles can be used multiple times for a single QRF.

### Please specify which Issue this PR Resolves.
closes #1307 

### How can the changes be tested?
See the previous QRF PR (#1200).