Added tool to log testing time (#1083)
## What type of PR is this.
1. [ ] Bug
2. [X] Enhancement

### What have you changed and why?
Information:
Added a tool which logs the time actively played on the server. This tool will be used by my test runs to ensure the time conditions are met. 

The tool gets started by
```[] spawn A3A_fnc_startTestingTimer```
and starts logging from then. On restart the script gets restarted automatically. There is no way to stop the script afterwards, unless you load a new save or execute
```testingTimerIsActive = false```
as server and do a save afterwards and restart the server.

The tool sleeps 5 minutes and then checks for all players (PvP ~sucks~ included) if they moved more than 5 meters from their last known position. For a max of 10 players the five minutes gets added to the timer, so 10 active players get counted as 50 minutes.

The output is done in hours to make it more readable.

Added a fix for the performance log to not spam the rpt if no player is online

### Please specify which Issue this PR Resolves.
None

### Please verify the following and ensure all checks are completed.

1. [X] Have you loaded the Mission in Singleplayer?
2. [ ] Have you loaded the Mission in a Dedicated Server?

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

********************************************************
Notes:
Havent checked the save/load functionality, worst case is the script does not start correctly and needs to be started manually