Revert aggro improvements (#1732)
Currently the most aggressive faction wins the map as their attacks are way more frequent to fight the player advancing. This however leads to the more fought faction (mostly Occupants) to stop the other faction (mostly Invaders).

Now there is a chance for the attacked faction to defend the attack and counter attack the attacking force, dependent on aggro ratio between them. Lower aggro increases chance of winning attacks/defending attacks.

In the end the faction which is mostly inactive, should take over the map quickly until they are actively fighting the rebels and therefor get active.