Doom move hq split (#1866)
## What type of PR is this.
1. [ ] Bug
2. [ ] Change
3. [X] Enhancement

### What have you changed and why?
Information:
Split the moveHQ function into two functions (moveHQ and canMoveHQ), where canMoveHQ contains all checks and returns the reasons if that is not possible. MoveHQ contains the steps to actually start the HQ moving process.

moveHQ has been updated with the new headers, replaced all tabs with spaces and restructured to code to a readable form.
    
There are no gameplay or mechanic changes. This is done for Dooms UI rework to work.

### Please specify which Issue this PR Resolves.
closes #1794

### Please verify the following and ensure all checks are completed.

1. [X] Have you loaded the mission in LAN host?
2. [ ] Have you loaded the mission on a dedicated server?

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Steps: 

********************************************************
Notes: There are still tabs in functions.hpp which should be removed, but this PR should stay as small as it is (will open issue)

* Added canMoveHQ function

* Updated moveHQ, added header, replaced tabs

* Added two comments

* Made canMoveHQ return all reasons instead of only one