Fix allowCrewInImmobile not being applied correctly to convoy vehicles (#2105)
### What type of PR is this.
1. [X] Bug
2. [ ] Change
3. [ ] Enhancement

### What have you changed and why?
allowCrewInImmobile (command that prevents crew from bailing out of an immobilized vehicle) was being applied to the convoy objective vehicle rather than the armoured vehicles due to a copypaste error. This PR fixes it so armoured vehicles are set correctly to allowCrewInImmobile.

### Please specify which Issue this PR Resolves.
closes #2104

### Is further testing or are further changes required?
1. [X] No
2. [ ] Yes (Please provide further detail below.)

### How can the changes be tested?
Easiest at high war level but quite difficult. Spawn a convoy with an armoured vehicle in it (tank, APC, AA, not MRAP) that's not the objective vehicle. Delete the other vehicles with zeus or whatever and then cripple the armoured vehicle. The crew should remain in the vehicle.

Probably not worth the testing effort.

Notes:
* Fix allowCrewInImmobile not being applied correctly to convoy vehicles

* Flatten vehArmor. Pretty sure that was a bug in the first place

* Fix third bug in convoy allowCrewInImmobile line